+++ packages/tables/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-tables@1.2.3...@zendeskgarden/react-tables@1.2.4) (2019-02-11)
+
+**Note:** Version bump only for package @zendeskgarden/react-tables
+
+
+
+
+
+## [1.2.3](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-tables@1.2.2...@zendeskgarden/react-tables@1.2.3) (2019-01-29)
+
+**Note:** Version bump only for package @zendeskgarden/react-tables
+
+
+
+
+
 ## [1.2.2](https://github.com/zendeskgarden/react-components/compare/@zendeskgarden/react-tables@1.2.1...@zendeskgarden/react-tables@1.2.2) (2019-01-10)
 
 **Note:** Version bump only for package @zendeskgarden/react-tables